reorganised the project structure making it more clear and concise
