+++ Middleware/.gitignore
@@ -5,7 +5,7 @@ env.json
 joularjx-result/*/all/total/methods/joularJX-*-all-methods-energy.csv
 joularjx-result/*/app/total/methods/joularJX-*-filtered-methods-energy.csv
 /citydata-files
-
+src/test/data/db/*
 HELP.md
 target/
 !.mvn/wrapper/maven-wrapper.jar
