+++ Middleware/src/main/java/ca/concordia/encs/citydata/core/controllers/ListProducerController.java
@@ -9,7 +9,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import ca.concordia.encs.citydata.core.contracts.Constants;
+import ca.concordia.encs.citydata.core.utils.Constants;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
